% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_fncs.R
\name{check_extracts_dir}
\alias{check_extracts_dir}
\title{Check for Subdirectory "Extracts" and Create If Necessary}
\usage{
check_extracts_dir()
}
\value{
nothing
}
\description{
Check for Subdirectory "Extracts" and Create If Necessary
}
