% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fncs.R
\name{create_data_tab}
\alias{create_data_tab}
\title{Apply a Function to an Array, For Extract Creation}
\usage{
create_data_tab(df, index, margin, tab.max, fnc, ...)
}
\arguments{
\item{index}{variable name in x to apply the specified function}

\item{margin}{variable name in x to stratify data in index}

\item{tab.max}{for a given extract, largest number of unique stratiying values}

\item{...}{optional arguments to fun}

\item{x}{a tibble}

\item{fun}{the function to be applied}
}
\value{
An array including results of function specified as fun, on variable
  specified as index, stratified by margin, first element is for entire
  array, remaining elements are for stratified
}
\description{
Apply a Function to an Array, For Extract Creation
}
\examples{
create_data_tab(nycflights13::flights, "air_time", "origin", 3, mean)
}
