% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codebook.R
\name{CreateCodebook}
\alias{CreateCodebook}
\title{Create a Dataset Codebook}
\usage{
CreateCodebook(df, df_file_name, df_title)
}
\description{
Create a codebook for a dataset including XXX for all variables.
}
