% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_extract.R
\name{CreateExtractMean}
\alias{CreateExtractMean}
\title{Create a Table Extract}
\usage{
CreateExtractMean(x, vars, tab_vars, notes, outputfile, info, stats = c(mean,
  median, min, max), test = NA, run_test = TRUE)
}
\arguments{
\item{x}{a tibble}

\item{vars}{a character vector include one or more variable name in x}

\item{tab_vars}{a character vector including one or more variable name}

\item{notes}{a character string including notes to include in the extract}

\item{outputfile}{a character string for which to write the extract}

\item{info}{hmm}

\item{stats}{statistical summaries to be included in the extract}

\item{test}{(optional) the name of a statistical test to be conducted hwen
run_test is set to TRUE, if left as NA the default test will be used}

\item{run_test}{(optional) should an appropriate statistical test be conducted}
}
\description{
A set of functions to create different types of table extracts. These extracts
are written to a comma-separated value (.csv) file.
}
\details{
There are several different types of table extracts that can be created
}
