% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_fncs.R
\name{get_max_unique}
\alias{get_max_unique}
\title{Maximum # of Unique Values in a List of Stratifying Variables}
\usage{
get_max_unique(x, strat_var)
}
\arguments{
\item{x}{a dataframe}

\item{strat_var}{character vector of a list of variables in x}
}
\value{
The number of unique values in variable in strat_var list of variables
  with largest number of unique values
}
\description{
Maximum # of Unique Values in a List of Stratifying Variables
}
