% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_fncs.R
\name{GetDateTime}
\alias{GetDateTime}
\title{Create a properly formatted date and time string}
\usage{
GetDateTime()
}
\value{
The current date and time
}
\description{
Create a properly formatted date and time string
}
