% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fncs.R
\name{length_nona}
\alias{length_nona}
\title{Length of an Object (Excluding Missing Values)}
\usage{
length_nona(x, na.rm = TRUE)
}
\arguments{
\item{x}{Array of numeric data.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
A number representing the number of non-missing values in an array.
}
\description{
Length of an Object (Excluding Missing Values)
}
\examples{
length_nona(nycflights13::flights$air_time)

### Compare this result to the following
length(nycflights13::flights$air_time)
}
