% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_summary.R
\name{meanSummary}
\alias{meanSummary}
\title{Functions used to add a mean value to codebooks}
\usage{
meanSummary(v, maxDecimals = 2)
}
\arguments{
\item{v}{Vector of values in a dataset}

\item{maxDecimals}{Maximum number of decimals to display (default = 2)}
}
\description{
A set of functions that will help to add the line of mean value
to codebooks.
}
