% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctZeroes.R
\name{pctZeroes}
\alias{pctZeroes}
\title{Functions used to add percentage of zeroes in variable to codebooks}
\usage{
pctZeroes(v, maxDecimals = 2)
}
\arguments{
\item{v}{Vector of values in a dataset}

\item{maxDecimals}{Maximum number of decimals to display (default = 2)}
}
\description{
A set of functions that will help to add the line of percentage of zero
values for a variable to codebooks.
}
