% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fncs.R
\name{q1}
\alias{q1}
\title{Return the First Quartile}
\usage{
q1(x, na.rm = TRUE)
}
\arguments{
\item{x}{Array of numeric data.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
A number representing the first quartile of the data in x.
}
\description{
Return the First Quartile
}
\examples{
q1(nycflights13::flights$distance)
}
