% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_fncs.R
\name{q3}
\alias{q3}
\title{Return the Third Quartile}
\usage{
q3(x, na.rm = TRUE)
}
\arguments{
\item{x}{Array of numeric data.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
A number representing the third quartile of the data in x.
}
\description{
Return the Third Quartile
}
\examples{
q3(nycflights13::flights$distance)
}
