% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableRExtracts.R
\docType{package}
\name{TableRExtracts}
\alias{TableRExtracts}
\alias{TableRExtracts-package}
\title{TableRExtracts: A package for creating statistical extracts}
\description{
The TableRExtracts package provides a set of useful functions for providing
documentation to a dataset and for creating statistical extracts of that
datset. These statistical extracts are written in .csv files and can be
useful for automatically populating tables for publication in Excel.
}
\section{TableRExtracts functions}{

}

